+++ sdk/data/azcosmos/cosmos_container.go
@@ -432,10 +432,10 @@ func (c *ContainerClient) ReadItem(
 	return response, err
 }
 
-// GetPartitionKeyRange retrieves all the distinct Partition Range keys froma a Cosmos container.
+// GetPartitionKeyRanges retrieves all the distinct Partition Range keys froma a Cosmos container.
 // ctx - The context for the request.
 // o - Options for the operation, can be nil.
-func (c *ContainerClient) GetPartitionKeyRange(ctx context.Context, o *PartitionKeyRangeOptions) (PartitionKeyRangeResponse, error) {
+func (c *ContainerClient) GetPartitionKeyRanges(ctx context.Context, o *PartitionKeyRangeOptions) (PartitionKeyRangeResponse, error) {
 	spanName, err := c.getSpanForContainer(operationTypeRead, resourceTypePartitionKeyRange, c.id)
 	if err != nil {
 		return PartitionKeyRangeResponse{}, err